Devex 864: only split the test files, not the code
