+++ shared/src/business/entities/cases/Case.attachIrsPractitioner.test.js
@@ -1,8 +1,8 @@
 const {
   applicationContext,
-} = require('../../../test/createTestApplicationContext');
-const { Case } = require('../Case');
-const { IrsPractitioner } = require('../../IrsPractitioner');
+} = require('../../test/createTestApplicationContext');
+const { Case } = require('./Case');
+const { IrsPractitioner } = require('../IrsPractitioner');
 
 describe('attachIrsPractitioner', () => {
   it('adds the user to the irsPractitioners', () => {
@@ -17,6 +17,7 @@ describe('attachIrsPractitioner', () => {
         userId: 'irsPractitioner',
       }),
     );
+
     expect(caseToVerify.irsPractitioners).not.toBeNull();
     expect(caseToVerify.irsPractitioners[0].userId).toEqual('irsPractitioner');
   });
